% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_when.R
\name{do_when}
\alias{do_when}
\title{Do when}
\usage{
do_when(.d, .w, .s)
}
\arguments{
\item{.d}{An expression to do upon a condition being met.}

\item{.w}{A condition when \code{.d} should be done.}

\item{.s}{Time in seconds to wait between checking for file existence.
The default is to wait 60 seconds between checking for a file.}
}
\value{
Output as generated by \code{.d}
}
\description{
Do something when something else is true
}
