% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_when_file.R
\name{do_when_file}
\alias{do_when_file}
\title{Do when a file exists}
\usage{
do_when_file(file_name, .f, wait_time = 60)
}
\arguments{
\item{wait_time}{Time in seconds to wait between checking for file existence.
The default is to wait 60 seconds between checking for a file.}

\item{file}{Name of file on which to trigger expression \code{.f}}

\item{expression}{A formula where the file name provided with \code{file} is
assigned as either ".x" or "file."}
}
\value{
The output of the supplied expression \code{.f}
}
\description{
Process an expression once a file exists.
}
