% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_when_time.R
\name{do_when_time}
\alias{do_when_time}
\title{Do when a time happens}
\usage{
do_when_time(.d, time, .s = 60)
}
\arguments{
\item{.d}{An expression to do upon a condition being met.}

\item{time}{A date-time specifying when \code{.d} should be done.}

\item{.s}{Time in seconds to wait between checking for file existence.
The default is to wait 60 seconds between checking for a file.}
}
\description{
Process an expression once a time is reached.
}
